Added location ability to articles, edited show views for User/Article to show location if applicable
