+++ app/views/users/new.html.erb
@@ -43,8 +43,8 @@
           </li>
           
           <li class="form-li">
-            <%= form.label 'Add Current Location to Profile', class: "checkbox" %>
-             <%= form.check_box :location, {}, 1, nil %>
+            <%= form.label :location, 'Add Current Location to Profile', class: "checkbox" %>
+            <%= form.check_box :location, {}, 1, nil %>
           </li>
 
           <li class="form-li li-submit">
