+++ db/migrate/20180627181238_add_location_to_articles.rb
@@ -1,0 +1,5 @@
+class AddLocationToArticles < ActiveRecord::Migration[5.2]
+  def change
+    add_column :articles, :location, :json
+  end
+end
